refactor: replace portfolio object with Portfolio class for tax calculations
