+++ capital-gains/src/calculator.ts
@@ -1,53 +1,41 @@
 import Decimal from "decimal.js";
 import { Operation, TaxResult } from "./types";
+import { Portfolio } from "./domain/entities/portfolio";
 import { Money } from "./domain/value-objects/money";
 
 Decimal.set({ precision: 20, rounding: Decimal.ROUND_HALF_UP });
 
 export function calculateTaxes(operations: Operation[]): TaxResult[] {
-  const portfolio = {
-    shares: 0,
-    averagePrice: new Decimal(0),
-    accumulatedLoss: new Decimal(0)
-  };
+  const portfolio = new Portfolio();
 
   return operations.map(operation => {
     if (operation.operation === 'buy') {
-      return processBuyHybrid(operation, portfolio);
+      return processBuyWithPortfolio(operation, portfolio);
     } else {
-      return processSellHybrid(operation, portfolio);
+      return processSellWithPortfolio(operation, portfolio);
     }
   });
 }
 
-function processBuyHybrid(operation: Operation, portfolio: { shares: number; averagePrice: Decimal; accumulatedLoss: Decimal; }): TaxResult {
-  const currentValue = Money.fromDecimal(portfolio.averagePrice).multiply(portfolio.shares);
-  const purchaseValue = Money.fromNumber(operation['unit-cost']).multiply(operation.quantity);
-  const newTotalShares = portfolio.shares + operation.quantity;
-
-  const newAverage = currentValue.add(purchaseValue).divide(newTotalShares);
-
-  portfolio.shares = newTotalShares;
-  portfolio.averagePrice = newAverage.toDecimal();
+function processBuyWithPortfolio(operation: Operation, portfolio: Portfolio): TaxResult {
+  const unitPrice = Money.fromNumber(operation['unit-cost']);
+  portfolio.recordPurchase(operation.quantity, unitPrice);
 
   return { tax: 0 };
 }
 
-function processSellHybrid(operation: Operation, portfolio: { shares: number; averagePrice: Decimal; accumulatedLoss: Decimal; }): TaxResult {
-  const salePrice = Money.fromNumber(operation['unit-cost']).multiply(operation.quantity);
-  const saleValue = salePrice.multiply(operation.quantity);
-  const costBasis = Money.fromDecimal(portfolio.averagePrice).multiply(operation.quantity);
+function processSellWithPortfolio(operation: Operation, portfolio: Portfolio): TaxResult {
+  const unitPrice = Money.fromNumber(operation['unit-cost']);
+  const quantity = operation.quantity;
+  const saleValue = unitPrice.multiply(quantity);
 
-  portfolio.shares -= operation.quantity;
-  if (portfolio.shares === 0) {
-    portfolio.averagePrice = new Decimal(0);
-  }
+  const costBasis = portfolio.calculateCostBasis(quantity);
+  const operationResult = saleValue.subtract(costBasis);
 
-  const result = saleValue.subtract(costBasis);
+  portfolio.recordSale(quantity);
 
-  if (result.isNegative() || result.isZero()) {
-    const loss = Money.fromDecimal(portfolio.accumulatedLoss).add(result.abs());
-    portfolio.accumulatedLoss = loss.toDecimal();
+  if (operationResult.isNegative() || operationResult.isZero()) {
+    portfolio.recordLoss(operationResult.abs());
 
     return { tax: 0 };
   }
@@ -57,19 +45,7 @@ function processSellHybrid(operation: Operation, portfolio: { shares: number; av
     return { tax: 0 };
   }
 
-  let netProfit = result;
-    const accumulatedLoss = Money.fromDecimal(portfolio.accumulatedLoss);
-
-  if (!accumulatedLoss.isZero()) {
-    if (netProfit.isGreaterThanOrEqual(accumulatedLoss)) {
-      netProfit = netProfit.subtract(accumulatedLoss);
-      portfolio.accumulatedLoss = new Decimal(0);
-    } else {
-      const remainingLoss = accumulatedLoss.subtract(netProfit);
-      portfolio.accumulatedLoss = remainingLoss.toDecimal();
-      netProfit = Money.zero();
-    }
-  }
+  const netProfit = portfolio.applyLossDeduction(operationResult);
 
   if (netProfit.isZero()) {
     return { tax: 0 };
@@ -82,4 +58,3 @@ function processSellHybrid(operation: Operation, portfolio: { shares: number; av
 }
 
 export default calculateTaxes;
-
